+++ psutil/_pslinux.py
@@ -81,7 +81,6 @@ HAS_CPU_AFFINITY = hasattr(cext, "proc_cpu_affinity_get")
 # Number of clock ticks per second
 CLOCK_TICKS = os.sysconf("SC_CLK_TCK")
 PAGESIZE = cext_posix.getpagesize()
-BOOT_TIME = None  # set later
 LITTLE_ENDIAN = sys.byteorder == 'little'
 UNSET = object()
 
@@ -1557,14 +1556,11 @@ def users():
 
 def boot_time():
     """Return the system boot time expressed in seconds since the epoch."""
-    global BOOT_TIME
     path = f"{get_procfs_path()}/stat"
     with open_binary(path) as f:
         for line in f:
             if line.startswith(b'btime'):
-                ret = float(line.strip().split()[1])
-                BOOT_TIME = ret
-                return ret
+                return float(line.strip().split()[1])
         msg = f"line 'btime' not found in {path}"
         raise RuntimeError(msg)
 
@@ -1899,19 +1895,20 @@ class Process:
 
     @wrap_exceptions
     def create_time(self, monotonic=False):
-        # Start time unit is expressed in jiffies (clock ticks per
-        # second). It never changes and is not affected by system clock
-        # updates.
+        # The 'starttime' field in /proc/[pid]/stat is expressed in
+        # jiffies (clock ticks per second), a relative value which
+        # represents the number of clock ticks that have passed since
+        # the system booted until the process was created. It never
+        # changes and is unaffected by system clock updates.
         if self._ctime is None:
             self._ctime = (
                 float(self._parse_stat_file()['create_time']) / CLOCK_TICKS
             )
         if monotonic:
             return self._ctime
-        # Add the uptime, returning seconds since the epoch (this is
-        # subject to system clock updates).
-        bt = BOOT_TIME or boot_time()
-        return self._ctime + bt
+        # Add the boot time, returning time expressed in seconds since
+        # the epoch. This is subject to system clock updates.
+        return self._ctime + boot_time()
 
     @wrap_exceptions
     def memory_info(self):